reorg code, make get metrics match pattern for tables, add to-geopandas methods, update examples and tests.
