+++ src/teehr/loading/location_crosswalks.py
@@ -9,7 +9,7 @@ from teehr.loading.duckdb_sql import (
 from teehr.loading.utils import (
     validate_dataset_structure,
 )
-from teehr.models.dataset.table_models import LocationCrosswalk
+from teehr.models.tables import LocationCrosswalk
 from teehr.loading.utils import merge_field_mappings
 import teehr.const as const
 import duckdb
