+++ src/teehr/models/str_enum.py
