+++ tests/setup_v0_3_study.py
@@ -2,7 +2,7 @@
 from pathlib import Path
 from teehr import Evaluation
 
-from teehr.models.dataset.table_models import (
+from teehr.models.tables import (
     Attribute
 )
